Avoid many more string allocations
Strings are now interned into a unified buffer, that is prepopulated
with the special calls that contain paths. The index can then
differentiate between them.

Also, boxes the interned str into a Rc<str> instead of a String so that
it can be shared between the hash-map and the vector of values.